+++ crates/validation_middleware_rust/src/lib.rs
@@ -1,146 +1,25 @@
+mod json;
+mod matcher;
+mod paths;
+mod sanitize;
+
+use json::{validate_json_bytes_streaming, walk_json_like};
+use matcher::{DangerousPatternMatcher, ValidationFailure, validate_string};
+use paths::{normalize_absolute_path, validate_resource_path_impl};
 use pyo3::prelude::*;
-use pyo3::types::{PyAny, PyDict, PyList, PyString};
+use pyo3::types::PyAny;
 use pyo3_stub_gen::define_stub_info_gatherer;
 use pyo3_stub_gen::derive::*;
-use regex::Regex;
-use serde::de::{self, DeserializeSeed, MapAccess, SeqAccess, Visitor};
-use std::fmt;
-use std::path::{Component, Path, PathBuf};
-
-const MAX_JSON_DEPTH: usize = 1024;
+use sanitize::sanitize_response_body_bytes;
+use std::path::PathBuf;
 
 #[gen_stub_pyclass]
 #[pyclass(name = "Validator")]
 pub struct CompiledValidator {
-    max_param_length: usize,
-    matcher: DangerousPatternMatcher,
-    allowed_roots: Vec<PathBuf>,
-    max_path_depth: usize,
-}
-
-struct DangerousPatternMatcher {
-    shell_metacharacters: bool,
-    path_traversal: bool,
-    control_characters: bool,
-    fallback_pattern: Option<Regex>,
-}
-
-impl DangerousPatternMatcher {
-    fn from_patterns(patterns: &[String]) -> PyResult<Self> {
-        let mut shell_metacharacters = false;
-        let mut path_traversal = false;
-        let mut control_characters = false;
-        let mut fallback_patterns = Vec::new();
-
-        for pattern in patterns {
-            match pattern.as_str() {
-                r"[;&|`$(){}\[\]<>]" => shell_metacharacters = true,
-                r"\.\.[\\/]" => path_traversal = true,
-                r"[\x00-\x1f\x7f-\x9f]" => control_characters = true,
-                _ => fallback_patterns.push(pattern.as_str()),
-            }
-        }
-
-        let fallback_pattern = if fallback_patterns.is_empty() {
-            None
-        } else {
-            let joined = fallback_patterns
-                .iter()
-                .map(|pattern| format!("(?:{pattern})"))
-                .collect::<Vec<_>>()
-                .join("|");
-            Some(Regex::new(&joined).map_err(|error| {
-                PyErr::new::<pyo3::exceptions::PyValueError, _>(error.to_string())
-            })?)
-        };
-
-        Ok(Self {
-            shell_metacharacters,
-            path_traversal,
-            control_characters,
-            fallback_pattern,
-        })
-    }
-
-    fn is_match(&self, value: &str) -> bool {
-        if value.is_ascii() {
-            let bytes = value.as_bytes();
-            if self.shell_metacharacters
-                && bytes.iter().any(|byte| {
-                    matches!(
-                        byte,
-                        b';' | b'&'
-                            | b'|'
-                            | b'`'
-                            | b'$'
-                            | b'('
-                            | b')'
-                            | b'{'
-                            | b'}'
-                            | b'['
-                            | b']'
-                            | b'<'
-                            | b'>'
-                    )
-                })
-            {
-                return true;
-            }
-            if self.path_traversal
-                && bytes
-                    .windows(3)
-                    .any(|window| window == b"../" || window == b"..\\")
-            {
-                return true;
-            }
-            if self.control_characters
-                && bytes
-                    .iter()
-                    .any(|byte| matches!(byte, 0x00..=0x1f | 0x7f..=0x9f))
-            {
-                return true;
-            }
-        } else {
-            if self.path_traversal {
-                let mut chars = value.chars();
-                let mut first = chars.next();
-                let mut second = chars.next();
-                for third in chars {
-                    if first == Some('.') && second == Some('.') && matches!(third, '/' | '\\') {
-                        return true;
-                    }
-                    first = second;
-                    second = Some(third);
-                }
-            }
-
-            for ch in value.chars() {
-                if self.shell_metacharacters
-                    && matches!(
-                        ch,
-                        ';' | '&' | '|' | '`' | '$' | '(' | ')' | '{' | '}' | '[' | ']' | '<' | '>'
-                    )
-                {
-                    return true;
-                }
-                if self.control_characters {
-                    let code = ch as u32;
-                    if matches!(code, 0x00..=0x1f | 0x7f..=0x9f) {
-                        return true;
-                    }
-                }
-            }
-        }
-
-        self.fallback_pattern
-            .as_ref()
-            .is_some_and(|pattern| pattern.is_match(value))
-    }
-}
-
-enum ValidationFailure {
-    MaxLength,
-    DangerousPattern,
+    pub(crate) max_param_length: usize,
+    pub(crate) matcher: DangerousPatternMatcher,
+    pub(crate) allowed_roots: Vec<PathBuf>,
+    pub(crate) max_path_depth: usize,
 }
 
 fn compile_validator(
@@ -161,458 +40,6 @@ fn compile_validator(
             .collect(),
         max_path_depth,
     })
-}
-
-fn validate_string(value: &str, validator: &CompiledValidator) -> Option<ValidationFailure> {
-    if value.is_ascii() {
-        if value.len() > validator.max_param_length {
-            return Some(ValidationFailure::MaxLength);
-        }
-    } else if value.chars().count() > validator.max_param_length {
-        return Some(ValidationFailure::MaxLength);
-    }
-
-    if validator.matcher.is_match(value) {
-        return Some(ValidationFailure::DangerousPattern);
-    }
-
-    None
-}
-
-enum StreamStop {
-    Failure(String, String),
-    MaxDepth,
-    InvalidJson(String),
-}
-
-const FAILURE_PREFIX: &str = "__validation_failure__:";
-const DEPTH_PREFIX: &str = "__validation_depth__";
-
-fn stream_stop_error<E>(stop: StreamStop) -> E
-where
-    E: de::Error,
-{
-    match stop {
-        StreamStop::Failure(key, error_type) => {
-            E::custom(format!("{FAILURE_PREFIX}{key}:{error_type}"))
-        }
-        StreamStop::MaxDepth => E::custom(DEPTH_PREFIX),
-        StreamStop::InvalidJson(message) => E::custom(message),
-    }
-}
-
-fn parse_stream_stop(error: serde_json::Error) -> StreamStop {
-    let message = error.to_string();
-    if let Some(rest) = message.strip_prefix(FAILURE_PREFIX) {
-        if let Some((key, error_type)) = rest.rsplit_once(':') {
-            let normalized_error_type = error_type
-                .split(" at line ")
-                .next()
-                .unwrap_or(error_type)
-                .to_owned();
-            return StreamStop::Failure(key.to_owned(), normalized_error_type);
-        }
-    }
-    if message.starts_with(DEPTH_PREFIX) {
-        return StreamStop::MaxDepth;
-    }
-    StreamStop::InvalidJson(message)
-}
-
-struct ValueSeed<'a> {
-    validator: &'a CompiledValidator,
-    depth: usize,
-    key_context: Option<&'a str>,
-    list_item_context: bool,
-}
-
-struct ValueVisitor<'a> {
-    seed: ValueSeed<'a>,
-}
-
-impl<'de, 'a> DeserializeSeed<'de> for ValueSeed<'a> {
-    type Value = ();
-
-    fn deserialize<D>(self, deserializer: D) -> Result<Self::Value, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        if self.depth > MAX_JSON_DEPTH {
-            return Err(stream_stop_error(StreamStop::MaxDepth));
-        }
-
-        deserializer.deserialize_any(ValueVisitor { seed: self })
-    }
-}
-
-impl<'de, 'a> Visitor<'de> for ValueVisitor<'a> {
-    type Value = ();
-
-    fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter.write_str("a JSON value")
-    }
-
-    fn visit_bool<E>(self, _value: bool) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_i64<E>(self, _value: i64) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_u64<E>(self, _value: u64) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_f64<E>(self, _value: f64) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_none<E>(self) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_unit<E>(self) -> Result<Self::Value, E> {
-        Ok(())
-    }
-
-    fn visit_str<E>(self, value: &str) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        if let Some(result) = validate_string(value, self.seed.validator) {
-            let key = if self.seed.list_item_context {
-                "list_item".to_owned()
-            } else {
-                self.seed.key_context.unwrap_or("payload").to_owned()
-            };
-            let error_type = match result {
-                ValidationFailure::MaxLength => "max_length".to_owned(),
-                ValidationFailure::DangerousPattern => "dangerous_pattern".to_owned(),
-            };
-            return Err(stream_stop_error(StreamStop::Failure(key, error_type)));
-        }
-
-        Ok(())
-    }
-
-    fn visit_borrowed_str<E>(self, value: &'de str) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        self.visit_str(value)
-    }
-
-    fn visit_string<E>(self, value: String) -> Result<Self::Value, E>
-    where
-        E: de::Error,
-    {
-        self.visit_str(&value)
-    }
-
-    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
-    where
-        A: SeqAccess<'de>,
-    {
-        let child_seed = ValueSeed {
-            validator: self.seed.validator,
-            depth: self.seed.depth + 1,
-            key_context: None,
-            list_item_context: true,
-        };
-
-        while seq.next_element_seed(child_seed.reborrow())?.is_some() {}
-        Ok(())
-    }
-
-    fn visit_map<A>(self, mut map: A) -> Result<Self::Value, A::Error>
-    where
-        A: MapAccess<'de>,
-    {
-        while let Some(key) = map.next_key::<String>()? {
-            map.next_value_seed(ValueSeed {
-                validator: self.seed.validator,
-                depth: self.seed.depth + 1,
-                key_context: Some(key.as_str()),
-                list_item_context: false,
-            })?;
-        }
-        Ok(())
-    }
-}
-
-impl<'a> ValueSeed<'a> {
-    fn reborrow<'b>(&'b self) -> ValueSeed<'b> {
-        ValueSeed {
-            validator: self.validator,
-            depth: self.depth,
-            key_context: self.key_context,
-            list_item_context: self.list_item_context,
-        }
-    }
-}
-
-fn validate_json_bytes_streaming(
-    raw_body: &[u8],
-    validator: &CompiledValidator,
-) -> PyResult<Option<(String, String)>> {
-    let mut deserializer = serde_json::Deserializer::from_slice(raw_body);
-    deserializer.disable_recursion_limit();
-    let streaming_result = {
-        let deserializer = serde_stacker::Deserializer::new(&mut deserializer);
-
-        (ValueSeed {
-            validator,
-            depth: 0,
-            key_context: None,
-            list_item_context: false,
-        })
-        .deserialize(deserializer)
-    };
-
-    match streaming_result {
-        Ok(()) => match deserializer.end() {
-            Ok(()) => Ok(None),
-            Err(error) => match parse_stream_stop(error) {
-                StreamStop::Failure(key, error_type) => Ok(Some((key, error_type))),
-                StreamStop::MaxDepth => Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-                    "JSON payload exceeds maximum supported nesting depth",
-                )),
-                StreamStop::InvalidJson(message) => {
-                    Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(format!(
-                        "Request body contains invalid JSON: {message}"
-                    )))
-                }
-            },
-        },
-        Err(error) => match parse_stream_stop(error) {
-            StreamStop::Failure(key, error_type) => Ok(Some((key, error_type))),
-            StreamStop::MaxDepth => Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-                "JSON payload exceeds maximum supported nesting depth",
-            )),
-            StreamStop::InvalidJson(message) => {
-                Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(format!(
-                    "Request body contains invalid JSON: {message}"
-                )))
-            }
-        },
-    }
-}
-
-fn walk_json_like(
-    py: Python<'_>,
-    data: &Bound<'_, PyAny>,
-    validator: &CompiledValidator,
-) -> PyResult<Option<(String, String)>> {
-    let mut stack = vec![(data.clone().unbind(), 0usize)];
-
-    while let Some((item, depth)) = stack.pop() {
-        if depth > MAX_JSON_DEPTH {
-            return Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-                "JSON payload exceeds maximum supported nesting depth",
-            ));
-        }
-
-        let bound_item = item.bind(py);
-
-        if let Ok(dict) = bound_item.cast::<PyDict>() {
-            for (key, value) in dict.iter() {
-                if let Ok(value_string) = value.cast::<PyString>() {
-                    if let Some(result) = validate_string(value_string.to_str()?, validator) {
-                        let key_string = key.str()?.to_string_lossy().into_owned();
-                        let error_type = match result {
-                            ValidationFailure::MaxLength => "max_length",
-                            ValidationFailure::DangerousPattern => "dangerous_pattern",
-                        };
-                        return Ok(Some((key_string, error_type.to_owned())));
-                    }
-                } else if value.is_instance_of::<PyDict>() || value.is_instance_of::<PyList>() {
-                    stack.push((value.unbind(), depth + 1));
-                }
-            }
-            continue;
-        }
-
-        if let Ok(list) = bound_item.cast::<PyList>() {
-            for child in list.iter().rev() {
-                if child.is_instance_of::<PyDict>() || child.is_instance_of::<PyList>() {
-                    stack.push((child.unbind(), depth + 1));
-                } else if let Ok(value_string) = child.cast::<PyString>() {
-                    if let Some(result) = validate_string(value_string.to_str()?, validator) {
-                        let error_type = match result {
-                            ValidationFailure::MaxLength => "max_length",
-                            ValidationFailure::DangerousPattern => "dangerous_pattern",
-                        };
-                        return Ok(Some(("list_item".to_owned(), error_type.to_owned())));
-                    }
-                }
-            }
-        }
-    }
-
-    Ok(None)
-}
-
-fn sanitize_response_body_bytes(body: &[u8]) -> Vec<u8> {
-    String::from_utf8_lossy(body)
-        .chars()
-        .filter(|ch| {
-            let code = *ch as u32;
-            !(matches!(code, 0x00..=0x08 | 0x0b | 0x0c | 0x0e..=0x1f | 0x7f..=0x9f))
-        })
-        .collect::<String>()
-        .into_bytes()
-}
-
-fn has_uri_scheme(path: &str) -> bool {
-    let Some((scheme, _rest)) = path.split_once("://") else {
-        return false;
-    };
-    let mut chars = scheme.chars();
-    let Some(first) = chars.next() else {
-        return false;
-    };
-    if !first.is_ascii_alphabetic() {
-        return false;
-    }
-    chars.all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '+' | '-' | '.'))
-}
-
-fn normalize_absolute_path(absolute: PathBuf) -> PathBuf {
-    let mut normalized = PathBuf::new();
-    for component in absolute.components() {
-        match component {
-            Component::Prefix(prefix) => normalized.push(prefix.as_os_str()),
-            Component::RootDir => normalized.push(Path::new("/")),
-            Component::CurDir => {}
-            Component::ParentDir => {
-                normalized.pop();
-            }
-            Component::Normal(part) => normalized.push(part),
-        }
-    }
-
-    normalize_windows_verbatim_prefix(normalized)
-}
-
-#[cfg(windows)]
-fn normalize_windows_verbatim_prefix(path: PathBuf) -> PathBuf {
-    let path_string = path.to_string_lossy();
-
-    if let Some(rest) = path_string.strip_prefix(r"\\?\UNC\") {
-        return PathBuf::from(format!(r"\\{rest}"));
-    }
-
-    if let Some(rest) = path_string.strip_prefix(r"\\?\") {
-        return PathBuf::from(rest);
-    }
-
-    path
-}
-
-#[cfg(not(windows))]
-fn normalize_windows_verbatim_prefix(path: PathBuf) -> PathBuf {
-    path
-}
-
-fn normalize_path(path: &str) -> Result<PathBuf, std::io::Error> {
-    let candidate = Path::new(path);
-    let absolute = if candidate.is_absolute() {
-        candidate.to_path_buf()
-    } else {
-        std::env::current_dir()?.join(candidate)
-    };
-
-    Ok(normalize_absolute_path(absolute))
-}
-
-fn resolve_absolute_path(path: PathBuf, symlink_depth: usize) -> Result<PathBuf, std::io::Error> {
-    if symlink_depth > 40 {
-        return Err(std::io::Error::new(
-            std::io::ErrorKind::InvalidInput,
-            "too many symlink expansions",
-        ));
-    }
-
-    let normalized = normalize_absolute_path(path);
-    let mut resolved = PathBuf::new();
-
-    for component in normalized.components() {
-        match component {
-            Component::Prefix(prefix) => resolved.push(prefix.as_os_str()),
-            Component::RootDir => resolved.push(Path::new("/")),
-            Component::CurDir => {}
-            Component::ParentDir => {
-                resolved.pop();
-            }
-            Component::Normal(part) => {
-                let candidate = resolved.join(part);
-                match std::fs::symlink_metadata(&candidate) {
-                    Ok(metadata) if metadata.file_type().is_symlink() => {
-                        let target = std::fs::read_link(&candidate)?;
-                        let target_path = if target.is_absolute() {
-                            target
-                        } else {
-                            resolved.join(target)
-                        };
-                        resolved = resolve_absolute_path(target_path, symlink_depth + 1)?;
-                    }
-                    Ok(_) => resolved.push(part),
-                    Err(error) if error.kind() == std::io::ErrorKind::NotFound => {
-                        resolved.push(part);
-                    }
-                    Err(error) => return Err(error),
-                }
-            }
-        }
-    }
-
-    Ok(normalize_absolute_path(resolved))
-}
-
-fn resolve_path(path: &str) -> Result<PathBuf, std::io::Error> {
-    resolve_absolute_path(normalize_path(path)?, 0)
-}
-
-fn validate_resource_path_impl(path: &str, validator: &CompiledValidator) -> PyResult<String> {
-    if has_uri_scheme(path) {
-        return Ok(path.to_owned());
-    }
-
-    if path.contains('\0') {
-        return Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-            "invalid_path: Invalid path",
-        ));
-    }
-
-    if path.contains("..") || path.contains("//") {
-        return Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-            "invalid_path: Path traversal detected",
-        ));
-    }
-
-    let resolved_path = resolve_path(path).map_err(|_| {
-        PyErr::new::<pyo3::exceptions::PyValueError, _>("invalid_path: Invalid path")
-    })?;
-
-    if resolved_path.components().count() > validator.max_path_depth {
-        return Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-            "invalid_path: Path too deep",
-        ));
-    }
-
-    if !validator.allowed_roots.is_empty()
-        && !validator
-            .allowed_roots
-            .iter()
-            .any(|root| resolved_path.starts_with(root))
-    {
-        return Err(PyErr::new::<pyo3::exceptions::PyValueError, _>(
-            "invalid_path: Path outside allowed roots",
-        ));
-    }
-
-    Ok(resolved_path.to_string_lossy().into_owned())
 }
 
 #[gen_stub_pymethods]
@@ -716,6 +143,7 @@ define_stub_info_gatherer!(stub_info);
 mod tests {
     use super::*;
     use pyo3::types::{PyDict, PyList};
+    use regex::Regex;
 
     #[test]
     fn validate_string_uses_character_count_not_utf8_bytes() {
